Refactor Xml PageSettings into a separate loader class
